Question 4 should have worked by now :(
Added the basis b and spin count N to spinspin_correlation, in order to
properly work in a subspace of H with Sz=const. The states I use to
verify it are now calculated using repeatedly applying the Hamiltonian
like in question 0. The correlation is 1 for x=0 and 0 otherwise. Crap.