Debugging commit
I think I found the bug and the solution: since in normalize_PEPS() I
call of any order of ind_i,ind_j, the resulting order of indices in
bondtensor() is flipped. The easiest solution for that that will also
be consistent with SVD_bondtensor() is to assume the first index to always
be the orthogonality center and flip the indices in normalize_PEPS()
if needed.